CLI utils moved out of ya-serv-api
