+++ core/serv-api/Cargo.toml
@@ -13,4 +13,5 @@ prettytable-rs = "0.8"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 serde_yaml = "0.8"
+ya-utils-cli = "0.1"
 url = "2.1.1"
