+++ utils/cli/src/lib.rs
@@ -1,0 +1,6 @@
+
+mod cmd;
+mod table;
+
+pub use cmd::CommandOutput;
+pub use table::ResponseTable;
